% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gantt.R
\name{create_gantt}
\alias{create_gantt}
\title{Generate Gantt chat}
\usage{
create_gantt(
  project,
  spots = NULL,
  by_date = FALSE,
  exact_date = FALSE,
  project_start_date = Sys.Date(),
  project_end_date = Sys.Date() + 365,
  colour_palette = wesanderson::wes_palette("Darjeeling1"),
  font_family = "sans",
  mark_quarters = FALSE,
  mark_years = FALSE,
  size_wp = 6,
  hide_wp = FALSE,
  size_activity = 4,
  size_text_relative = 1,
  label_wrap = FALSE,
  month_number_label = TRUE,
  month_date_label = TRUE,
  x_axis_position = "top",
  colour_stripe = "lightgray",
  alpha_wp = 1,
  alpha_activity = 1,
  line_end = "round",
  month_breaks = 1,
  show_vertical_lines = TRUE,
  axis_text_align = "right"
)
}
\description{
Generate Gantt chat
}
\examples{
\dontrun{
}
}
\author{
Phil J. Bouchet
}
