% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.narwsim.R
\name{expand.narwsim}
\alias{expand.narwsim}
\title{Posterior}
\usage{
\method{expand}{narwsim}(
  obj,
  nsamples = 10000,
  xrange = c(0.05, 0.6),
  prop.sd = c(0.01, 0.01),
  gauss.approx = FALSE,
  rw.start = c(0.05, 0.05),
  max.iter = 500,
  ntune = 500,
  plot = TRUE,
  plot.n = 200,
  ...
)
}
\description{
Posterior sampling
}
\examples{
\dontrun{
library(narwind)

animals <- run_model(10)
plot(animals)
}

------------------------------------------------------------
------------------------------------------------------------
------------------------------------------------------------
------------------------------------------------------------
------------------------------------------------------------
------------------------------------------------------------
------------------------------------------------------------
------------------------------------------------------------
}
\author{
Phil J. Bouchet
}
