% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_mh.r
\name{gam_mh}
\alias{gam_mh}
\title{Posterior}
\usage{
gam_mh(
  b,
  ns = 10000,
  burn = 1000,
  t.df = 40,
  rw.scale = 0.25,
  thin = 1,
  pb = TRUE,
  verbose = FALSE
)
}
\description{
Posterior sampling
}
\examples{
\dontrun{
library(narwind)

animals <- run_model(10)
plot(animals)
}

}
\author{
Phil J. Bouchet
}
