% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_vessels.R
\name{map_vessels}
\alias{map_vessels}
\title{Vessel traffic and strike risk}
\usage{
map_vessels(
  obj = NULL,
  z = "dist",
  strike_scalar = 1e-07,
  which.month = 1:12,
  speed = NULL,
  baseline = TRUE,
  gantt = FALSE,
  do.plot = FALSE,
  spgdf = FALSE
)
}
\description{
Vessel traffic and strike risk
}
\examples{
\dontrun{
}
...............................
...............................
...............................
...............................
...............................
...............................
}
\author{
Phil J. Bouchet
}
