% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.narwproj.R
\name{summary.narwproj}
\alias{summary.narwproj}
\title{Summary}
\usage{
\method{summary}{narwproj}(obj, ...)
}
\description{
Summary information
}
\examples{
\dontrun{
library(narwind)

animals <- run_model(10)
plot(animals)
}
}
\author{
Phil J. Bouchet
}
